+++ backend/crates/eiffelvis_gen/src/base_event.rs
@@ -10,7 +10,7 @@ pub struct BaseMeta {
     #[serde(rename = "type")]
     pub event_type: String,
     pub version: String,
-    pub time: u64,
+    pub time: u128,
 }
 
 #[derive(Default, Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Deserialize)]
